+++ element.go
@@ -24,8 +24,13 @@ const (
 	encodingPrefixUncompressed = 0x04
 )
 
-// errParamInvalidPointEncoding indicates an invalid point encoding has been provided.
-var errParamInvalidPointEncoding = errors.New("invalid point encoding")
+var (
+	// ErrParamInvalidPointEncoding indicates an invalid point encoding has been provided.
+	ErrParamInvalidPointEncoding = errors.New("invalid point encoding")
+
+	// ErrParamNilElement indicates a forbidden nil Element.
+	ErrParamNilElement = errors.New("nil element")
+)
 
 // Element implements the Element interface for the secp256k1 group element.
 type Element struct {
@@ -89,8 +94,13 @@ func (e *Element) Identity() *Element {
 }
 
 // Add sets the receiver to the sum of the input and the receiver, and returns the receiver.
+// If element is nil, Add panics.
 func (e *Element) Add(element *Element) *Element {
-	return e.add(element)
+	if element == nil {
+		panic(ErrParamNilElement)
+	}
+
+	return e.addProjectiveComplete(e, element)
 }
 
 // Double sets the receiver to its double, and returns it.
@@ -108,21 +118,22 @@ func (e *Element) Negate() *Element {
 }
 
 // Subtract subtracts the input from the receiver, and returns the receiver.
+// If element is nil, Subtract panics.
 func (e *Element) Subtract(element *Element) *Element {
 	if element == nil {
-		return e
+		panic(ErrParamNilElement)
 	}
 
 	q := element.copy().negate()
 
-	return e.add(q)
+	return e.addProjectiveComplete(e, q)
 }
 
 // Multiply sets e to the Scalar multiplication with the given Scalar in constant time,
-// and returns e.
+// and returns e. If scalar is nil, Multiply panics.
 func (e *Element) Multiply(scalar *Scalar) *Element {
 	if scalar == nil {
-		return e.Identity()
+		panic(ErrParamNilScalar)
 	}
 
 	r0 := newElement()
@@ -132,7 +143,7 @@ func (e *Element) Multiply(scalar *Scalar) *Element {
 	for i := 255; i >= 0; i-- {
 		bit := uint64(bits[i])
 		cSwapElements(bit, r0, r1)
-		r1.Add(r0)
+		r1.addProjectiveComplete(r1, r0)
 		r0.Double()
 		cSwapElements(bit, r0, r1)
 	}
@@ -143,7 +154,12 @@ func (e *Element) Multiply(scalar *Scalar) *Element {
 }
 
 // Equal returns 1 if the elements are equivalent, and 0 otherwise.
+// If element is nil, Equal panics.
 func (e *Element) Equal(element *Element) int {
+	if element == nil {
+		panic(ErrParamNilElement)
+	}
+
 	return e.isEqual(element)
 }
 
@@ -153,7 +169,12 @@ func (e *Element) IsIdentity() bool {
 }
 
 // Set sets the receiver to the value of the argument, and returns the receiver.
+// If element is nil, Set panics.
 func (e *Element) Set(element *Element) *Element {
+	if element == nil {
+		panic(ErrParamNilElement)
+	}
+
 	return e.set(element)
 }
 
@@ -201,19 +222,19 @@ func Secp256Polynomial(y, x *field.Element) {
 func (e *Element) DecodeCoordinates(x, y [32]byte) error {
 	fex, reduced := field.New().FromBytesWithReduce(x)
 	if reduced == 0 {
-		return errParamInvalidPointEncoding
+		return ErrParamInvalidPointEncoding
 	}
 
 	fey, reduced := field.New().FromBytesWithReduce(y)
 	if reduced == 0 {
-		return errParamInvalidPointEncoding
+		return ErrParamInvalidPointEncoding
 	}
 
 	var y2 field.Element
 	Secp256Polynomial(&y2, fex)
 
 	if y2.Equals(field.New().Square(fey)) != 1 {
-		return errParamInvalidPointEncoding
+		return ErrParamInvalidPointEncoding
 	}
 
 	e.x.Set(fex)
@@ -227,11 +248,11 @@ func (e *Element) DecodeCoordinates(x, y [32]byte) error {
 // on failure.
 func (e *Element) DecodeCompressed(data []byte) error {
 	if len(data) != elementLengthCompressed {
-		return errParamInvalidPointEncoding
+		return ErrParamInvalidPointEncoding
 	}
 
 	if data[0] != encodingPrefixEven && data[0] != encodingPrefixOdd {
-		return errParamInvalidPointEncoding
+		return ErrParamInvalidPointEncoding
 	}
 
 	/*
@@ -243,7 +264,7 @@ func (e *Element) DecodeCompressed(data []byte) error {
 	// Set x in the field, and return an error if it's not reduced.
 	x := field.New()
 	if _, reduced := x.FromBytesWithReduce([32]byte(data[1:])); reduced == 0 {
-		return errParamInvalidPointEncoding
+		return ErrParamInvalidPointEncoding
 	}
 
 	var y2 field.Element
@@ -252,7 +273,7 @@ func (e *Element) DecodeCompressed(data []byte) error {
 	y, isSquare := field.New().SqrtRatio(&y2, field.New().One())
 	if isSquare != 1 {
 		// Point is not on curve
-		return errParamInvalidPointEncoding
+		return ErrParamInvalidPointEncoding
 	}
 
 	cond := y.Sgn0() ^ uint64(data[0]&1)
@@ -269,11 +290,11 @@ func (e *Element) DecodeCompressed(data []byte) error {
 // on failure.
 func (e *Element) DecodeUncompressed(data []byte) error {
 	if len(data) != elementLengthUncompressed {
-		return errParamInvalidPointEncoding
+		return ErrParamInvalidPointEncoding
 	}
 
 	if data[0] != encodingPrefixUncompressed {
-		return errParamInvalidPointEncoding
+		return ErrParamInvalidPointEncoding
 	}
 
 	return e.DecodeCoordinates([32]byte(data[1:33]), [32]byte(data[33:]))
@@ -284,7 +305,7 @@ func (e *Element) Decode(data []byte) error {
 	switch len(data) {
 	case elementLengthIdentity:
 		if data[0] != encodingPrefixIdentity {
-			return errParamInvalidPointEncoding
+			return ErrParamInvalidPointEncoding
 		}
 
 		e.Identity()
@@ -295,7 +316,7 @@ func (e *Element) Decode(data []byte) error {
 	case elementLengthUncompressed:
 		return e.DecodeUncompressed(data)
 	default:
-		return errParamInvalidPointEncoding
+		return ErrParamInvalidPointEncoding
 	}
 }
 
@@ -517,12 +538,4 @@ func (e *Element) addProjectiveComplete(u, v *Element) *Element {
 	e.z.Set(z3)
 
 	return e
-}
-
-func (e *Element) add(element *Element) *Element {
-	if element == nil {
-		return e
-	}
-
-	return e.addProjectiveComplete(e, element)
 }